re-organize pkg
