+++ pkg/addon/istio_metric_test.go
@@ -1,4 +1,4 @@
-package prometheus
+package addon
 
 import (
 	"fmt"
